pyproject.toml
[tool.poetry]
name = "learning_site"
version = "0.1.0"
description = "The website for the sign language learning tool."
authors = ["ArthurVanRemoortel <arthurvanremoortel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
gunicorn = "^20.1.0"
python-dotenv = "^0.21.0"
Django = "^4.1.1"
libsass = "^0.21.0"
django-compressor = "^4.1"
django-sass-processor = "^1.2.1"
Pillow = "^9.2.0"
numpy = "^1.23.4"
ffmpeg-python = "^0.2.0"
keras = "^2.10.0"
scikit-video = "^1.1.11"
tensorflow = "^2.10.0"
opencv-python = "^4.6.0.66"
sklearn = "^0.0"
pandas = "^1.5.0"
psycopg2 = "^2.9.5"
djangorestframework = "^3.14.0"
django-filter = "^22.1"
PyYAML = "^6.0"
aiosmtpd = "^1.4.2"
mediapipe = "^0.9.0.1"
scikit-learn = "^1.2.0"
tqdm = "^4.64.1"
seaborn = "^0.12.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
